Redesigned the Graph Iterator and bugfixes for edges with None vertices:
The way the iterator on the Graph object worked was a bit convoluted and
unnecessarily so. Now it is more intuitive in its function and can deal
with None values on Edges.

With the edges there were some bugs left over from switching to allow
None edges connected to Edges.