Fix deleting Vertices leaving references to the deleted Vertex on Edges:
As it stood using the Vertex.delete_from() method would at times leave
behind references to the deleted vertex in the Edges connected to the
Vertex. To fix this the Production.apply() function needed to be changed
because it implicitly relied on this behaviour to switch connections of
edges from a deleted Vertex to a newly added Vertex.