Fix matplotlibs NavigationToolbar throwing GTK Warnings:
The NavigationToolbar on my matplotlib figures was throwing GTK+
negative content heigth warnings. I realized that this happened only if
the NavigationToolbar was in some way contained within a wx.Notebook.
Switching out the wx.Notebook for a aui.AuiNotebook got rid of all the
Warnings.

The test1.py file was used in debugging this error, it contains a
minimal example showcasing this bug.