Added saving and automatic calculation of x/y coordinates for elements:
All nodes in a graph now save their x and y coordinates in a attribute
(called, surprisingly, 'x' and 'y' respectively). For new elements added
within a production the new coordinates are calculated based on the
barycenter of the mapped elements between mother and daughter graph. If
this calculation does not produce the expected results, it is possible
to manually define a calculation for the x and y coordinates by defining
'new_x' or 'new_y' attributes in a daughter element.