First version of apply to oldest generations first condition:
This condition allows one to specify that a particular production
should be applied to the oldest generation it matches. Right now the
calculation of what set of generations of elements is considered the
oldest favours having a single element from the lowest generation.

I will next explore using the average generation of elements to see
if this produces better results.