Support position calculations in productions with vector operations:
I have added a set of simple vector operations that ease the creation
of new rules for the calculation of position of new elements. They are
much easier to use than the bare-bones attribute access.

What is missing right now is the ability to define vectors the
calculations have access to within the gui, it can only be done in the
yaml file.