Made it possible to match graphs while retaining geometric relations:
By geometric relation I mean the x,y positioning as in this element is
left of another element. Now a production can specify the
'.geometric_order' condition with a value of 'True' and then the
production will only be applied to matches which keep the geometric
ordering between elements the same.