Improvement to automatic element position with directed edges:
Up until now directed edges where ignored during the calculation of
new positions, this could lead to elements being added on the wrong
side of a directed edge.

With this update, if there is a directed edge present in the mother
graph of a production, then this specific edge will be used to
calculate the degree by which to rotate new elements.